04_flask_books_library/requirements.txt
Faker==15.3.4
Flask==2.2.2
flask-migrate==4.0.0
flask-wtf==1.0.1
python-dotenv==0.21.0
